chore(new-hope): migrate main.js for storybook@latest
